Prevent ship placement if any square is already occupied
The code that passes the test has 2 switch cases, it's not elegant, but
it does work. I could try to simplify the accounting for orientation
when trying to place a ship. I could also use a try/catch situation to
place the ships? Would it not place the ship if it didn't work?